Triangle. Refs #25.

+++ triangle/Triangle.java
@@ -3,7 +3,7 @@ import java.util.Scanner;
 public class Triangle{
     public static void main(String[] args){
         Scanner scanner = new Scanner(System.in);
-        int base, height, ans;
+        double base, height, ans;
         System.out.println("What is the base of a triangle? ");
         base = scanner.nextInt();
         System.out.println("What is the height of a triangle? ");
